+++ demo/app/src/main/java/com/loginradius/demo/MainActivity.java
@@ -61,6 +61,5 @@ public class MainActivity extends AppCompatActivity {
         btnlogin.setOnClickListener(listener);
         btnregister.setOnClickListener(listener);
         btnforgotpassword.setOnClickListener(listener);
-
 }
 }